Renamed ConcurrentTest to Benchmark, ConcurrentTestRunner to BenchmarkRunner and ConcurrentTransactionTest to TransactionBenchmark. Move all files to foam.nanos.bench from foam.nanos.test
