+++ src/foam/nanos/nanos.js
@@ -35,6 +35,7 @@ FOAM_FILES([
   { name: "foam/nanos/auth/token/AbstractTokenService" },    
   { name: "foam/nanos/auth/ChangePasswordView"},
   { name: "foam/nanos/auth/User" },
+  { name: "foam/nanos/bench/Benchmark" },
   { name: "foam/nanos/boot/NSpec" },
   { name: "foam/nanos/session/Session" },
   { name: "foam/nanos/client/ClientBuilder" },
@@ -53,7 +54,6 @@ FOAM_FILES([
   { name: "foam/nanos/script/Script" },
   { name: "foam/nanos/test/Test" },
   { name: "foam/nanos/test/TestBorder" },
-  { name: 'foam/nanos/test/ConcurrentTest' },
   { name: "foam/nanos/cron/Cron" },
   { name: "foam/nanos/export/ExportDriverRegistry"},
   { name: "foam/nanos/export/ExportDriver" },
