+++ tools/classes.js
@@ -106,6 +106,7 @@ var classes = [
   'foam.nanos.actioncommand.ActionCommand',
   'foam.nanos.NanoService',
   'foam.nanos.boot.NSpec',
+  'foam.nanos.bench.Benchmark',
   'foam.nanos.auth.EnabledAware',
   'foam.nanos.auth.Group',
   'foam.nanos.auth.LastModifiedAware',
@@ -149,7 +150,6 @@ var classes = [
   'foam.nanos.notification.push.FirebasePushService',
   'foam.nanos.script.Script',
   'foam.nanos.test.Test',
-  'foam.nanos.test.ConcurrentTest',
   'foam.nanos.cron.Cron',
   'foam.dao.history.PropertyUpdate',
   'foam.dao.history.HistoryRecord',
